Make thread-based feature state observing optional, with worker process-based feature state observing the default
Threaded-based observing is not supported by django and should not be used in production, but can be useful for development and tests.\n-Currently, all tests succeed without any observer functionality at all. We should try to maintain this orthogonality.\n-The main purpose of the observer here is to handle incomplete client disconnections.